[fix](planner)only forbid literal value in AnalyticExpr's order by list (#21820)
pick from master https://github.com/apache/doris/pull/21819

The original error is the window function's order by clause doesn't support a literal value. But we use Expr's isConstant method to check if it's an literal, this is wrong. So the pr uses the correct method( isLiteral ) to check if an expr is literal to solve this problem

```sql
select
    sortNum,
    BITMAP_UNION_COUNT (c.pv) over (ORDER BY **sortNum** ) totalNum
from(
select 
    ifnull(a.sortNum, b.sortNum) sortNum,
    BITMAP_UNION (ifnull(a.c_pv, b.c_pv)) pv
from
    (select **_1 sortNum_**, c_pv from ${tableName2} t where t.c_int = 1) a
full join
    (select **_2 sortNum_**, c_pv from ${tableName2} t where t.c_int = 2) b
    on a.sortNum = b.sortNum
GROUP BY
    sortNum
ORDER BY
    sortNum
```