[Fix]Fix insert select missing audit log when connect follower FE (#36454)
## Proposed changes
Fix when a ```insert select``` is executed in Follower, audit log could
missing query statistics.
This is because ```audit log``` is logged in the connect FE, but request
is forward to master FE, then the coord FE is master FE, BE report query
statistics to cood FE, finally the connected Follower could not get
reported query statistics, audit log missing query statistics.
We can add a new field to mark client connected FE, then be report query
statistics to the connected FE.
Besides, I do refactor for FE's WorkloadRuntimeStatusMgr to make logic
more clear and add some log in be.