[cherry-pick](branch-2.0)  Pick "[Enhancement](merge-on-write) Support  dynamic delete bitmap cache (#32991)" (#37602)
The default delete bitmap cache is set to 100MB, which can be
insufficient and cause performance issues when the amount of user data
is large. To mitigate the problem of an inadequate cache, we will take
the larger of 5% of the total memory and 100MB as the delete bitmap
cache size.

## Proposed changes

Issue Number: close #xxx

<!--Describe your changes.-->